Remove unused FieldInfo (#1540)
