+++ be/src/olap/tablet_meta.cpp
@@ -123,7 +123,7 @@ TabletMeta::TabletMeta(int64_t table_id, int64_t partition_id,
             column->set_precision(tcolumn.column_type.precision);
             column->set_frac(tcolumn.column_type.scale);
         }
-        uint32_t length = FieldInfo::get_field_length_by_type(
+        uint32_t length = TabletColumn::get_field_length_by_type(
                 tcolumn.column_type.type, tcolumn.column_type.len);
                 column->set_length(length);
         column->set_index_length(length);
