+++ be/test/olap/CMakeLists.txt
@@ -42,7 +42,7 @@ ADD_BE_TEST(skiplist_test)
 ADD_BE_TEST(delta_writer_test)
 ADD_BE_TEST(serialize_test)
 ADD_BE_TEST(olap_meta_test)
-ADD_BE_TEST(field_info_test)
+ADD_BE_TEST(decimal12_test)
 ADD_BE_TEST(storage_types_test)
 ADD_BE_TEST(aggregate_func_test)
 ADD_BE_TEST(rowset/segment_v2/bitshuffle_page_test)
